requirements.txt
aiohttp>=3.7.4
faiss-cpu>=1.7.2
nltk>=3.6.2
numpy>=1.21.0
openai>=0.27.0
rank_bm25>=0.2.2
requests>=2.25.1
sentence-transformers>=2.2.0
torch>=1.9.0
tqdm>=4.62.0
transformers>=4.10.0